functional routing /templates and /projects
