+++ apps/builder/app/dashboard/projects/index.ts
@@ -1,1 +1,2 @@
 export { Projects } from "./projects";
+export { Templates } from "./templates";
