+++ apps/builder/app/dashboard/projects/projects.tsx
@@ -2,19 +2,17 @@ import { Flex, Grid, Text, rawTheme } from "@webstudio-is/design-system";
 import type { DashboardProject } from "@webstudio-is/dashboard";
 import { EmptyState } from "./empty-state";
 import { Panel } from "../shared/panel";
-import { ProjectCard, ProjectTemplateCard } from "./project-card";
+import { ProjectCard } from "./project-card";
 import { CreateProject } from "./project-dialogs";
 
 type ProjectsProps = {
   projects: Array<DashboardProject>;
-  projectTemplates: Array<DashboardProject>;
   hasProPlan: boolean;
   publisherHost: string;
 };
 
 export const Projects = ({
   projects,
-  projectTemplates,
   hasProPlan,
   publisherHost,
 }: ProjectsProps) => {
@@ -46,32 +44,6 @@ export const Projects = ({
           })}
         </Grid>
       </Flex>
-
-      {projectTemplates.length > 0 && (
-        <Flex direction="column" gap="3">
-          <Flex justify="between">
-            <Text variant="brandSectionTitle" as="h2">
-              Templates
-            </Text>
-          </Flex>
-          <Grid
-            gap="6"
-            css={{
-              gridTemplateColumns: `repeat(auto-fill, minmax(${rawTheme.spacing[31]}, 1fr))`,
-            }}
-          >
-            {projectTemplates.map((project) => {
-              return (
-                <ProjectTemplateCard
-                  project={project}
-                  publisherHost={publisherHost}
-                  key={project.id}
-                />
-              );
-            })}
-          </Grid>
-        </Flex>
-      )}
     </Panel>
   );
 };
